Refactor code to put "main" package on root and package "dvara" in "lib" folder
