+++ .travis.yml
@@ -3,5 +3,6 @@ env: GO_RUN_LONG_TEST=1
 go:
   - 1.2
   - 1.3
+  - tip
 install:
   - go get -t ./...
