+++ lib/dvara/doc.go
