+++ lib/dvara/protocol_test.go
