+++ lib/dvara/proxy_test.go
