+++ lib/dvara/response_rewriter_test.go
