+++ lib/dvara/rs_state.go
