+++ lib/dvara/rs_state_test.go
